\docType{data}
\name{airports.spain}
\alias{airports.spain}
\title{Spanish airports statistical data}
\format{A dataframe with 43 observations on the following 16 variables.

\describe{

\item{ICAO_code}{Airport ICAO code}
\item{name}{Airport title}
\item{lon}{Airport longitude}
\item{lat}{Airport latitude}
\item{year}{Year of observation; 2010}
\item{APM}{A number of carried passengers}
\item{ATM}{A number of of air transport movements served by an airport}
\item{freight}{A total volume of freight served by an airport}
\item{revenue}{A total airport revenue}
\item{DA}{A total deprecation/amortization costs of airport operation}
\item{staff_cost}{A total staff cost of airport operation}
\item{runways}{A number of airport runways}
\item{terminals}{A number of airport terminals}
\item{WLU}{A total volume of work load units (WLU) served by an airport}

}}
\source{
\describe{ \item{}{TDC (2012). Informe de fiscalizacion de
la imputacion por la entidad "Aeropuertos Espanoles y
Navegacion Aerea" (AENA) a cada uno de los aeropuertos de
los ingresos, gastos, e inversiones correspondientes a la
actividad aeroportuaria, en los ejercicios 2009 y 2010.,
Tribunal de Cuentas, Spain, Doc 938.} \item{}{Eurostat
(2013). European Statistics Database, Statistical Office of
the European Communities (Eurostat)} \item{}{Airports'
statistical reports (2011)} }
}
\description{
The \code{spfrontier} package includes the dataset
\code{airports}, containing information about Spanish
airports infrastructure and traffic statistics in 2010.
}

