\docType{data}
\name{airports}
\alias{airports}
\title{European airports statistical data}
\format{A dataframe with 113 observations on the following 10 variables.

\describe{

\item{ICAO_code}{Airport ICAO code}
\item{name}{Airport title}
\item{lon}{Airport longitude}
\item{lat}{Airport latitude}
\item{PAX}{A number of carried passengers}
\item{ATM}{A number of of air transport movements served by an airport}
\item{freight}{A total volume of freight served by an airport}
\item{runways}{A number of airport runways}
\item{checkins}{A number of airport check-in facilities}
\item{gates}{A number of airport gates}

}}
\source{
\describe{ \item{}{Eurostat (2013). European Statistics
Database, Statistical Office of the European Communities
(Eurostat)} \item{}{Airports' statistical reports(2011)} }
}
\description{
The \code{spfrontier} package includes the dataset
\code{airports}, containing information about European
airports infrastructure and traffic statistics in 2011.
}

