\name{spfrontier-package}
\alias{spfrontier-package}
\docType{package}
\title{
Spatial Stochastic Frontier estimation
}
\description{
Spatial Stochastic Frontier estimation
}
\details{
\tabular{ll}{
Package: \tab spfrontier\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-10-07\cr
License: \tab LGPL\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Dmitry Pavlyuk <Dmitry.V.Pavlyuk@gmail.com>
}
\references{
Anselin, L. (1988). Spatial Econometrics: Methods and Models. Kluwer Academic Publishers, Dordrecht, The Netherlands.
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
}