\docType{package}
\name{spfrontier-package}
\alias{spfrontier-package}
\title{Spatial Stochastic Frontier}
\description{
Spatial Stochastic Frontier
}
\details{
A set of tools for estimation (MLE) of various spatial
specifications of stochastic frontier models
}
\author{
Dmitry Pavlyuk \email{Dmitry.V.Pavlyuk@gmail.com}
}
\keyword{frontier}
\keyword{spatial}
\keyword{stochastic}

