\name{spfrontier-package}
\alias{spfrontier-package}
\alias{spfrontier}
\docType{package}
\title{
Spatial Stochastic Frontier estiomation
}
\description{
Spatial Stochastic Frontier estiomation
}
\details{
\tabular{ll}{
Package: \tab spfrontier\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-10-07\cr
License: \tab LGPL\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Dmitry Pavlyuk <Dmitry.Pavlyuk@tsi.lv>
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}